libs/ocr_core/pyproject.toml
[project]
name = "ocr-core"
version = "0.1.0"
description = "Shared OCR domain + pipeline"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7.0",
    "pillow>=10.4.0",
    "numpy>=2.0.0",
    "pyyaml>=6.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python-headless>=4.8.0",
    "vietocr",
    "craft-text-detector",
]

[tool.uv.sources]
vietocr = { path = "../vietocr" }
craft-text-detector = { path = "../craft-text-detector" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ocr_core"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["ocr_core"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
